+++ src-docs/src/views/form_controls/field_number.js
@@ -31,6 +31,7 @@ export default class extends Component {
           placeholder="Placeholder text"
           value={this.state.value}
           onChange={this.onChange}
+          aria-label="Use aria labels when no actual label is in use"
         />
 
         <EuiSpacer size="m" />
@@ -40,6 +41,7 @@ export default class extends Component {
           value={this.state.value}
           onChange={this.onChange}
           disabled
+          aria-label="Use aria labels when no actual label is in use"
         />
 
         <EuiSpacer size="m" />
@@ -49,6 +51,7 @@ export default class extends Component {
           value={this.state.value}
           onChange={this.onChange}
           isLoading
+          aria-label="Use aria labels when no actual label is in use"
         />
 
         <EuiSpacer size="m" />
@@ -59,6 +62,7 @@ export default class extends Component {
           onChange={this.onChange}
           disabled
           isLoading
+          aria-label="Use aria labels when no actual label is in use"
         />
 
         <EuiSpacer size="m" />
@@ -68,6 +72,7 @@ export default class extends Component {
           value={this.state.value}
           onChange={this.onChange}
           readOnly
+          aria-label="Use aria labels when no actual label is in use"
         />
       </Fragment>
     );