+++ src-docs/src/views/form_controls/field_password.js
@@ -30,6 +30,7 @@ export default class extends Component {
           placeholder="Placeholder text"
           value={this.state.value}
           onChange={this.onChange}
+          aria-label="Use aria labels when no actual label is in use"
         />
 
         <EuiSpacer size="m" />
@@ -39,6 +40,7 @@ export default class extends Component {
           value={this.state.value}
           onChange={this.onChange}
           disabled
+          aria-label="Use aria labels when no actual label is in use"
         />
 
         <EuiSpacer size="m" />
@@ -48,6 +50,7 @@ export default class extends Component {
           value={this.state.value}
           onChange={this.onChange}
           isLoading
+          aria-label="Use aria labels when no actual label is in use"
         />
 
         <EuiSpacer size="m" />
@@ -58,6 +61,7 @@ export default class extends Component {
           onChange={this.onChange}
           isLoading
           disabled
+          aria-label="Use aria labels when no actual label is in use"
         />
       </Fragment>
     );