+++ src-docs/src/views/form_controls/range.js
@@ -34,6 +34,7 @@ export default class extends Component {
           max={200}
           value={this.state.value}
           onChange={this.onChange}
+          aria-label="Use aria labels when no actual label is in use"
         />
 
         <EuiSpacer size="m" />
@@ -45,6 +46,7 @@ export default class extends Component {
           value={this.state.value}
           onChange={this.onChange}
           disabled
+          aria-label="Use aria labels when no actual label is in use"
         />
       </Fragment>
     );