+++ src-docs/src/views/form_controls/select.js
@@ -36,6 +36,7 @@ export default class extends Component {
           options={this.options}
           value={this.state.value}
           onChange={this.onChange}
+          aria-label="Use aria labels when no actual label is in use"
         />
 
         <EuiSpacer size="m" />
@@ -45,6 +46,7 @@ export default class extends Component {
           value={this.state.value}
           onChange={this.onChange}
           disabled
+          aria-label="Use aria labels when no actual label is in use"
         />
 
         <EuiSpacer size="m" />
@@ -54,6 +56,7 @@ export default class extends Component {
           value={this.state.value}
           onChange={this.onChange}
           isLoading
+          aria-label="Use aria labels when no actual label is in use"
         />
 
         <EuiSpacer size="m" />
@@ -64,6 +67,7 @@ export default class extends Component {
           onChange={this.onChange}
           isLoading
           disabled
+          aria-label="Use aria labels when no actual label is in use"
         />
       </Fragment>
     );